feat: styling for image collection
